requirements.txt
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instaloader==4.14.1
pillow==11.3.0
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
python-telegram-bot==21.10
pytube==15.0.0
requests==2.32.3
sniffio==1.3.1
urllib3==2.4.0
yt-dlp==2025.7.21
